Improved Wall serialization
Now it uses a (slightly) smarter method, which takes advantage of the tendency for walls to exist in groups (it's not particularly clever, so only groups in the y dimension really matter).  Also, I did away with SparseMapLayer because the slight improvement in space efficiency probably doesn't make up for the time loss or code complexity.  More headers get precompiled now.