Changed Switchable behavior
Now when a Switchable is linked to multiple Signalers, it will behave "properly" (it will be "active" if and only if at least one Signaler is active) - before, linking a Switchable to multiple Signalers was essentially undefined behavior, as the undo system didn't even support it (and it would have been fairly easy to crash the game before this). See multi_sig_gate_test.map.